Unschedule validate_mutlipath from zfcp testsuite
The module validate_multipath has been failing in the zfcp testsuite,
due to changes in the fcp infrastructure on the z/VM hypervisor.
After the chages we have two disks attached to the system (an extra 200
GB disk has been added), which results in two WWIDs being displayed in
the multipath config.
This causes the module to fail because the `$wwid` parameter is supposed
to only contain one value.