Remove unavailable python2 module from hpc_autoyast
The sle-module-python2 is not available as module in sle15sp4.
I checked the installed system and python2 comes with the base system.

in this respect the autoyast schema doesnt need the particular sections and they are going to exclude
from the installation in this case. For all the other versions stays with the sle-module-python2

To accomplish this the xml is converted to a template which allows embedded perl code.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>