Containers: simpler output validation for container-suseconnect-zypp lm
For some reason, the following lines belong to stderr:
2022/01/10 19:10:16 Installed product: SLES-15.3-x86_64
2022/01/10 19:10:16 Registration server set to https://scc.suse.com

And the following line is already regular stdout:
All available modules:

This output validation works normally, but it fails on VMWare,
HyperV and Xen, I guess due to how the serial output is treated
in those backends.

We just need the string "All available modules" to be validated.