Wait for Installation Settings screen to be loaded
validate_ssh_service_enabled test module checked that Installation
Screen is loaded by ensuring that widget with "id=proposal" is shown on
the screen. Then 'assert_true' is called immediately and it failed,
because the content of 'proposal' widget was not loaded.

The commit ensures that the the content is loaded before validating that
the port is opened.

Related issue: https://progress.opensuse.org/issues/104703