Add allow vendor change for cases with packagehub and smt cases
When performing migration with packagehub in addon list, it would ask
whether agree to change vendor from SUSE to openSUSE for some packages,
which is reasonable because the packages from packagehub are openSUSE
packages; For smt registration type migration cases, see bsc#1120169
Comment#4, packagehub will be added to the target modules list. So we
add --allow-vendor-change for these cases.