Do not press return when installation error occurs
When the 'An error occured during the installation.' OK popup occurs,
sending return (and as a result hitting the OK button) will result in
all ttys logging out.
This results in post_fail_hooks failing and logs not being saved,
because any console accessed from that point onward has already
been logged out.
So sending return should be avoided in the case of the installation
error OK popup, in order for the post_fail_hooks to not fail.