Make use of is_shown to show what is done in SUT.
In order to make sure we are on the right page, we implemented for each
page a function method called is_shown. Then in the controller, we have
methods like "get_user_creation_page" which always call the "is_shown"
method, and each method from the page is using "get_user_creation_page"
hence each time calling is_shown. This would be in my opinion a good
place to show "what is shown".