Sanitize variable obtained from SUT with script_output
The variable $robot_test in the sles4sap/sys_param_check test module is
assigned with the output of a command in SUT executed with script_output().
Even though this works without major issues when BACKEND=qemu, it is
possible that in different backends strings in the variable can come with
unexpected characters.

This commit sanitizes the string, by removing CR & LF.