Fix erratic behavior of windows network setup
This fix uses the windows from wsl/boot_windows as base.
Due to the issue that the network adapter is sometimes (possibly
depending on the worker setup?) named "Ethernet" and sometimes
"Ethernet 2" there is the need to first find the name and use that
to actually set up the network.
This fix implements that.
For that, powershell cmdlets are used instead of netsh.