Implement hardware tests for RPi for SLE JeOS
SLE doesn't do DHCP and SSH with a default pw (like openSUSE does)
so this test needs to rely on a preprocessed image.
This image is generated by openQA using qemu backend
via create_hdd_rpi.yaml.

The SUT RPi is connected to an as sd card multiplexer
which is also connected to the worker via USB, so the worker
can flash that image onto the sdcard and then switch
the sdcard back to the SUT.
The SUT RPi is power cycled from the worker using a wifi power plug.

The hardware test is done via jeos-realhw.yaml.

The worker (which is a RPi as well) spans up a wifi network
and is visible via bluetooth to test the RF functions
of the SUT.

Ticket: https://progress.opensuse.org/issues/105852