Upload basic logs prior to the full problem detection logs
basic_logs contains loadavg - we want to get this 'as soon as possible'
to the point where the error occurred. If we get this only after
collecting all other logs, a few minutes have passed, which has a
great chance to make the numbers appear much lower than what is
interesting for us.