Use die_on_timeout=>0 in lib/sles4sap when required
PR#14027 introduces a change in script_run() that makes it fail whenever
a timeout with the SUT occurs, modifying the previous behavior when the
call would have returned undef. When the previous behavior is expected,
the call must include die_on_timeout => 0. This commit fixes the one
call to script_run() in lib/sles4sap that relies on an undef return
value.