nvme_checks: Do not use 'nmve show-regs' in smoke-test scenario
'nvme show-regs' accesses the raw registers via mapping IO space to
userspace. Today most distros enable CONFIG_IO_STRICT_DEVMEM which
prevents this operation. This can be overwritten by the kernel command
line option 'iomem=relaxed'. But this is not recommended for
production systems.

Hence using 'nvme show-regs' in a smoke-test for production images is
not particular useful. There are other nvme-cli commands which qualify
for smoke-tests but not this one. Remove it.

Signed-off-by: Daniel Wagner <dwagner@suse.de>