Fix yast2_lan_restart for SLE 15 SP1
The test is used in maintenance job group and was false positive before
the build 20220122-1.

The test expected that the network will be reloaded which is incorrect,
as NO changes were made in the config.

This commit updates the test to expect that no reload or restart of
network is happened as this seems to be fixed by the update.

Related ticket: https://progress.opensuse.org/issues/105358