Adjust ibtest to run against science:hpc repo
install SCIENCE_HPC_REPO with higher priority if it is defined. For now i use the SLE_15_SP3 repo which is available. TODO: fix this when SLE_15_SP4 is published
Add the IBTEST_TIMEOUT. This like likely do not make any big difference but the default is kinda strict as the /ibtest/ might exceed this limit.
Set it to 7200 has enough time until the test reach its own limit of the OpenQA jobs.

In addition, the assert_script_run causes issues as it is waiting too long for the serial return code and i am sure that the test doesnt really need to be asserted in this case.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>