Improve libyui REST API timeout before MAX_JOB_TIME
Once we hit MAX_JOB_TIME is game over, there is not chance to run code in post_fail_hooks by design.
As the process of performing installation requires a short interval to check due to sometimes we need to stop the reboot timeout in less than 10 seconds, creating a screenshot for each check is not possible, it would be too many of them.

What we can do is to avoid reaching that MAX_JOB_TIME waiting for the widget to exist only because we use a wait_until inside another wait_until. By passing timeout=0 to the methods which are executed in the inner block we can exit earlier from this long wait and get a proper error before hitting the maximum time set for the job. Also we measure the time instead of using a counter.

Additionally we can avoid to double the sleep time that we introduced for safety at the beginning of development with libyui REST API for this particular cases with zero timeout. For other cases this delay should be there, otherwise it fails due to some screen are not completed loaded and we don't check it properly in early code with new libyui REST API development.