Build a kernel on sle15sp4
This PR corrects steps required to build a kernel on sle15sp4.

The keys for the signing are not available, so there are two solutions.
One is to recreate them and set the proper value on /CONFIG_MODULE_SIG_KEY/ or disable the feature.
I think for this test case it is enough to go with the later.

I dont know how this used to work before but we need to copy the kernel to the boot and update the bootloader before the reboot.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>