Refactor add_suseconnect_product with retries
This method is used in multiple places, and we have observed that
it fails randomly due to network issues (timeouts, SCC not reachable,
etc). In most cases, restarting the test helps (unless there is a
major outage that tages long to be fixed).

This code does a similar thing what script_retry does, but with a
bit of more logic and info messages.