Prevent change of Apparmor selection when is presented in Patterns
the Installer of 15SP4 requires Apparmor pattern activation by default.
Remove of the unnecessary variable /LINUX_SECURITY_MODULE/. and enable /set_linux_security_to_none/ only when it is missing from the patterns.
In such way, the /set_linux_security_to_none/ will not run when the Apparmor is already selected and there wouldnt be a problem with the installation.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>