Adapt yast2_kdump after UI change
It seems like there was some UI change that makes it necessary to hit enter twice: we enter 320,
which is more than the maximum allowed value. So Yast changes it to the max allowed value, 287,
after we type enter, but then we need to type enter again. Another option would be to set
$memory_kdump to 287 only for aarch64, but that might cause problems again if the value changes
for some reason. Typing enter twice does not hurt in this case.