Integrate boost library into mpi_master
Modify the `mpi_master` in order to run HPC scientific libraries. The beginning is with [boost](https://theboostcpplibraries.com/boost.mpi-simple-data-exchange)
The sample of the source code exercises the exchange data between two processes. the process with rank 0 receives data with recv().
The process with rank 1 sends data with send(). If you start the program with more than two processes, the other processes exit without doing anything.
According to this, only two nodes are required. That means that the code can be run in a single node and even if the module uses `--host` with three nodes including the master,
only two of them will be chosen to run the binary. We can identify and check those, using the mpirun paramenter `-print-all-exitcodes`. This will report the node and the exit code of its execution.

`get_mpi_src` is used to fetch the source code and the appropiate compiler for that particular source code. This is accomplished with `HPC_LIB`. if the variable is not set
the other MPI implementations will be used(mpich, openmpi, mvapich2)

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>