JeOS: help troubleshoot btrfs_autocompletion issues
There are some not-so-sporadic issues in this module for JeOS tests
that fail when doing zypper install bash-completion with the error:
System management is locked by the application with pid XYZ...

It's difficult to reproduce the issue as it appears randomly.

Printing the list of processes before running zypper might help
to spot the issue when it happens again.