Add scipy HPC library test
Install `python3-scipy-gnu-hpc` and run a /hello/ python script.
To do so i use a python module /mpi4py/ which helps in translating MPI syntax
for C++ to Python. The script is really simple and it just validates that the
availability of the scipy(in fact numpy comes along too).
TODO: Update script with a more realistic scenario

After the installation some env variables are not populated so a
logout/restart is required, such as some init scripts will
rerun in order to set those variables. (/usr/share/lmod/lmod/init/bash,
/etc/profile.d, etc). After relogin variables as `MODULEPATH` should be
available in the env and `module av` should be able to find the modules
installed from `*-gnu-hpc` like _scipy_.

I introduce a new module, `formatter.pm`. The aim is to prepare the MPI
command for execution. This is needed for simplicity of the various expressions
of the commands against the several MPI implementations in combination the
numerous scientific modules. The idea is to be able to express as following
for maximux flexibility:

- $mpi_str_obj->all_nodes('path/to/bin');
- $mpi_str_obj->single_host('python path/to/bin');

I consider that the `formatter` is not fully ready but it seems to work for
now.

I have some additional concerns for `setup_scientific_module`. I imagined it
as one place to have all the config steps for each HPC library. However i
found that there would be some which will be available from /spack/ in the
future and are not available in sle15sp4. One of them is *boost*. that
particular block is commented out as the steps will not work for the time
being.

Finally it is just confusing to have this line inside `relogin_root` but sshd
is inactive after login back and needs restart.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>