[HPC] Remove blocks of code refers to boost library
As per https://www.suse.com/releasenotes/x86_64/SLE-HPC/15-SP4/index.html#_removed_packages
the code relating to `boost` is not accurate. The only value of the code is that it is testing
the compilation with the *g++*.

However, i havent removed the line in the `get_mpi_src` which defines the source code to be used,
as it is convinience for now for this purpose. However a better approach should be consider and this
code will have to changed/removed in the near future.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>