Ignore false softfail bsc#1191112 on normal system installation
The softfail pointed to https://bugzilla.suse.com/show_bug.cgi?id=1191112 is
not reproducable on a normal installation. With this commit the
`PUBLISH_HDD_1` is used to distiguish those jobs and keep the softfail only
for the absolute problematic jobs.

This approach seems the most convenience compare to others. Which most of them
will required much more changes and maintance. The main problem is that i
didnt found an easy way to detect when the issue occurs. With needling would
have to do too many checks and maintain a batch of new unnecessary needles
just for a workaround. Also the condition block doesnt look like it can be
moved anywhere else, such as being specific in cases which the problem
appears.

Another solution would be to remove or replace the `record_soft_failure` with
`record_info` but as there is not detection will keep shown up in the jobs
even if they will not be reported as softfailed jobs in the results.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>