Append condition which produce bsc#1191112 on HPC
There are some jobs which do not contain `PUBLISH_HDD_1` on HPC group
which still perform full installation, however. I appended the
condition to check if the sle_product is HPC.

This will not prevent the false softfail to showed up in all the other
groups. This is only to maintain the status report on HPC.

Revert if you find other HPC products failing because of this.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>