Implement boost module test installed with spack
The steps are very similar to `mpi_*` modules, with the difference of
installing different packages which mainly differ the install path and the
use of environment modules.

The main aim in this PR is to initialize Spack testing which installs and
loads boost module and run a hello_world. Testing includes other spack
features like uninstall and unload modules.

The steps in the module has been testing only on sle15sp4. The plan is to
merge those modules with the mpi_*. So the modules will subject to some
refactoring to provide a better design and integrate more modules to test.

Also the jobs uses 1G memory. Note that in the manual testing this reproduced
some issues, and as such if the jobs start showing defects during the
compilation by spack, it is recommended to increase QEMURAM to 4.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>