wicked: Fix duplicate mac-addr errors in wicked-testsuite
For some tests, e.g. macvtap or bridge-with-dummy, we use predefined
fixed mac-addresses.
If the order of the interfaces is such, that the bridge gets the
mac-address of the dummy-interface, while we have the same job parallel
running on a other multi-machine setup, the backend ovs-switch has two
times the same mac-address. The result are unexpected network errors.