Perform different checks if HANA volumes are mounted via NFS
The sles4sap/hana_test test module currently expects the HANA
installation to have been performed on local disks by either
sles4sap/hana_install or sles4sap/wizard_hana_install. This limits the
scenarios on which the module can be used to test a HANA installation.

This commit adds code to sles4sap/hana_test so it is able to
determine when the HANA volumes are mounted over NFS, and perform
checks different than the LVM commands on the vg_hana volume groupi as
it currently does.

Additionally, a new SKIP_HANADB_QUERY setting is introduced to skip
the hdbsql checks, for those scenarios where the HANA SYSTEM password
is not available.