Add a schedule to install openSUSE Tumbleweed
Some tests require hardware that was purchased for QA of SLES, and thus
we cannot simply connect that hardware to openSUSE's openQA instance.

Maybe sharing hardware will be possible in the future, but for now we
just want to be able to install and test Tumbleweed on our internal
hardware, too.

This adds a simple schedule to install Tumbleweed to our baremetal
machines using ipxe_install.

Signed-off-by: Michael Moese <mmoese@suse.com>