BCI: make use of markers to execute the tests
So far, BCI tests are using all the images of the same family in the
tests, e.g. dotnet environment is pulling all the dotnet image
versions available for dotnet.
By providing this marker using "-k bci/dotnet-aspnet_6.0" we make
sure the tests are only executed for this specific image.
This way we avoid running the tests for all the images in the same
family and save a lot of time.

This PR also adds running common tests in test_all.py