Configure nfs sharing between hpc master and slave nodes
Trying to simulate a real-world hpc setup, where the compute nodes are bare
machines which merely used to perform computation. This also makes the jobs a
bit faster as we do not need to provision those nodes with hpc packages. The
only compromise here is some required dependencies which hpc modules install
and are required to run mpi binaries.

To do so we need to export the location of the binary of which will run to test
mpi and the location of the installed HPC modules.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>