Remove pipe in mpi commands wherever is possible
Looking for a way to execute a script_run which would retrieve the output and
the exit code in one step, i failed. Using `script_output` doesnt work exactly
in the direction which was intented, as it ignores the exit code of the
execution. So for now, the most direct and easy solution is to remove the
pipes wherever is possible. With this, no more false positives are expected
and the logs can be found in the _serial_terminal.txt_ anyway. Exception is
the block which acts upon mvapich2.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>