Prevent timeout on aarch64 when mpirun invoke with mapich2
For some reason mapich2 takes longer time in the latest builds on older
sle15 versions and timeouts on aarch64. mapich2 encounter a known issue
with a segmentation fault and the test should check and raise a soft fail
tho. Increase the timeout of script_run to avoid the test to fail.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>