Reorganize test
Create a new schedule to create a jumphost HDD image
Create a new init test, designed to be executed on the Trento jumphost
HDD image. It only takes care to refresh and eventually switch branch in
the css-trento repo code.
split test in two files: one generic and one for the cypress tests.