Prevent known issues on HPC break the mvapich2 tests
Some known bugs for older SLE versions are breaking the tests but there is no
intention to be fixed or they are already having a fix in newer versions.
I follow the same practice which already work in the module using grep but
maybe i have to think a better approach in the future.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>