Provide user-virtio-terminal console option
The `select_serial_terminal` picks by default `root-virtio-terminal` but the
subroutine provides parameter to select a console as normal user.

Somehow this found misconfigured. The condition was like that
`$console = $root ? 'root-virtio-terminal' : 'virtio-terminal';`. There was
two main issues. One was that the propor console was not added during the
`init_consoles` and the second was that the variable was not a valid option.

The first is solved easily by adding another line to call `add_console`.
In the second case the correct name is `user-virtio-terminal` instead
`virtio-terminal`. the underline variable is evaluated later against a regex
to assign some other variables, one of them is *$user* and another is *type*
which each are required for the user _login_ and the console _initialization_.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>