Create get_compute_nodes_deps function on hpcbase library
Refactor mpi_slave.pm and spack_slave.pm moving common lines in the hpcbase
library. This particular block of code is used to select which packages are
needed to be installed in the compute nodes which use nfs to mount the HPC
modules.

This commit increases the timeout of some commands as a resolution of timeout
failures when it runs against aarch64 which spack performs really slow.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>