Refactor and expand Infiniband tests preparation
In preparation for running Infiniband Tests also on openSUSE, we need to
add some more preparation code, so we should split the preparations into
their own module. This helps keeping the test code a bit cleaner.
So this moves most of the preparation needed for master and slave jobs
into a new ibtests_prepare.pm.
Additionally, the schedules also need to contain this new module.

Signed-off-by: Michael Moese <mmoese@suse.com>