Use normal user to run HPC relative tasks on mpi jobs
Once the `user-virtio-console` resolved we can now switch easily to the normal
user for all the mpi implementations. As per `select_serial_terminal` was
called earlier to get a root console, we have to kill that session, because
the implementation doesnt taking care of this scenario and fails in a login
check.
During the implementation i encounter some timeouts and such i bump up the
particular parameter in specific invocations. However not sure whether it was
a sporadic issue on my local OpenQA or something else.
Finally i modified the `generate_and_distribute_ssh` to make copies only on
the compute nodes.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>