Use normal user to run HPC relative tasks on spack jobs
To run MPI binaries on the compute nodes as a regular user the management node
has to export the `/opt/spack` and `$HOME/spack` in additional of the HPC
libraries.

Initialy i thought that sharing the whole `$HOME` could work but during the
testing that failed. Likely because it is override previous steps that we have
performed, for instance the ssh key configuration. Anyhow, the functions which
setup the NFS were slightly modified to accept a hash ref with the absolute
paths to export. That impacted also the *mpi_master* ans *mpi_slave* modules.

Due to the need to run some commands as root, there are several switches
between the root and the regural user, which look ungly and need to improve in
the future.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>