Containers: use new registry from Amazon ECR
The VM where the current registry service is running today will be
decommissioned soon as we are migrating the Public Cloud accounts.
I found a solution using directly the ECR service instead of creating
a VM which is prone to failure and requires maintenance.
ECR provides a public registry for our openQA tests.