Use Public AWS ECR registry
The VM where the current registry service is running today will be
decommissioned soon as we are migrating the Public Cloud accounts.

There is solution using the public ECR Gallery which basically contains
the same images of DockerHub.
https://gallery.ecr.aws

There are some limitations when pulling images with docker, like
toomanyrequests: Rate exceeded
but with using retries, we can workaround the problem.

This solution is desired as we don't need to maintain any registry
ourselves which has caused some issues in the past.

https://progress.opensuse.org/issues/98415