Fix yast_user by adding the IP of the NIS server
yast_user failed because if ypwhich returning an IP.
This change adds IP of the nis server through a variable.
ypwhich is a tool that "return[s] name of NIS server or map master".
but it sometimes returns the IP address instead of the hostname, so the test failed
because of this.
The change also re-enables yast_user, introduces the variable
"NAMESERVER", and fixes wrong formatting in variable.md

See https://progress.opensuse.org/issues/112271