Run slurm configuration merely against compute nodes
The operations which try to copy the ssh keys and the slurm configuration
consider all the slave(compute) nodes including the master(management) one.
However i think this is not necessary. In this commit i change this to run
merely against the compute nodes.

For some reason the behavior of those operations started failing, when the
where trying to copy data from master to itself. I didnt dig into it but this
shouldnt be like that anyway IMO. Anyway merging this seems to resolve the
problem.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>