Redesign the migration regression cases
We redesign the migration regression cases that spliting a long case into
several cases: after migration, publsh a qcow, and then do regression
tests with the qcow; for the regression tests, we use INCLUDE_MODULES
setting to load the regression test modules. The reason that fix the
service check is we check the upgraded services on the second case, it can
not check the result in the first job.