GNOME: assert focus on xterm
As it is described in [[leap][qe-core] test fails in xterm - mistyping test string](https://progress.opensuse.org/issues/111752)
there are several typing issues observed in `xterm` usage. Either the
test string is not typed correctly or the focus of the `xterm` window is
lost. This PR should handle focus issue.

- [opensuse-15.4-DVD-Backports-Incidents-x86_64-Build20220712-3-gnome@64bit-2G](https://openqa.opensuse.org/tests/2464982#step/updates_packagekit_gpk/23)