Try to mitigate "VNC typing issues" with disabled key repeat
Independent of fixing the actual performance problems I think it can
help in general for openQA tests if we configure the systems we test in
with a disabled key repeat to mitigate the issues, i.e. "xset -r"
within at least X11 sessions.

This commit is adding a new method in the "x11test" base class and is
using that in a dedicated module that should be scheduled right after we
ensured the desktop runner to be responsive and usable.

Related progress issue: https://progress.opensuse.org/issues/110542