ALP: create new product "alp" with its main.pm
This ALP PoC in openQA can have its own product directory,
similar to microos, leap-micro or sle-micro.
For these jobs, we will be using DISTRI=alp, which means
new way of scheduling tests.
I am proposing to create this main.pm which will be the main
entry point for scheduling tests for ALP. Most of them are
the same as for MicroOS/Leap-Micro, and I might come up with
a common schedule file file for ALL these transactional-server
based products, but for now, I would prefer to keep it with
single main schedule.

Related ticket: https://progress.opensuse.org/issues/112409