Fix slurm config and munge distribution
Distribution of the keys and config files should include the replica master
node, which was neglected with previous commit. Ommitting that, caused a
series of issues. For starter, slurmctld service on the backup node couldnt
start, and without the munge key the node could not authenticate in the
cluster.

`FastSchedule` was removed as it is not needed and a warning is writen in the
logs.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>