Restart cups service after configuration changes
As long as cups daemon is running prior to this test case, starting the
service becomes a noop therefore configuration changes are not applied.

Warning message: `lpadmin: Raw queues are deprecated and will stop
working in a future version of CUPS.` is not handled in this PR as it is
harmless in the current version of `cups`.