lib/x11utils: ensure_unlocked_screen: Avoid waiting after screen change
Currently it presses Esc to show the lock screen, waits for the lock screen
to appear (through wait_screen_change) and then it does a wait_still_screen
which can take long enough for the lock screen to disappear again.

Avoid the latter by skipping straight to the next match after a screen change
was detected.