Create smoke test for HPC trilinos package
As per comment https://progress.opensuse.org/issues/112790#note-3 the only
requirement is a smoke test for trilinos. I quote the comment also here for
future reference:
Trilinos has been removed from SLE15 SP4. There are no new products which will
have this package. It may still appear in maintenance cases, but for this, we
will only need a smoke test - ie an installation test. Trilinos itself would
be extremely cumbersome to test as it consists of a wide range of individual
libraries - each of which would have to be tested.

So this test would be enable for maintainance prior to SLE15 SP4 just for the
rare updating cases.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>