Create autoyast profiles and schedulers for HPC
ay profiles are generated with the clone.pm module and were adjusted in order
to remove unnecessary elements and modified to run in each sle15 version.
I assume that the sle12 might be a bit different.

After created and tested 4 different xmls i merged the ones of the same arch.
So each arch should work for either the textmode or gnome_textmode. I used the
SSC_ADDONS variable which in the `*_gnome_textmode` seems to be a unique
identifier. But not sure if that the optimal approach. Instead we could use
another job variable, with best candidate `DESKTOP`. But this is not defined in
the test and for now i just didnt touch anything in the settings to make it work.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>