Uninstall updated packages after patch
If there are multiple patches, each patch will work with not updated packages
Each patch is showing L2 & L3 updated packages info, if package was updated
in revious patch then the info will be soft failed because there is no change
This will be improved with following changes

-rework patches list, to drop duplicate entries, e.g. happning with LTSS SAP
-some packages can have multiple versions installed, thus store last or first
 record depending if looking for old or new version
-don't add src packages into the list
-disable update repo and do zypper dup to get back previous package versions
 unfortunatelly this can't handle all packages, so manually remove updated
 package if there are more than one versions installed

There are still some pacakges which is hard to downgrade back to previous
version, like rpm. It will still happen L2 & L3 info will be not accurate when
there are multiple pathes, it's the best I could do.