Delete drawboard function from WebSocketsTest
Due to poo#114532, in the sake of stability, that we can safely
delete the function that tests the drawboard example.
We already cover the rest of Tomcat's WebSocket examples,
and, in reality, this drawboard example is not really tested
with multiple connections as it would ideally be the case.

At the same time, increase the timeout value for some tomcat
test modules to reduce the impact by performance.