Collect logs from parallel jobs when one node is failing
The solution to make multi-machine jobs to collect logs was prevented mainly
by the fact of the barrier's usage. When the parent failed the barriers
following the failed command is never released.

In this patch this is what is changing:
- Collect some basic logs after main tests are done. This is for keeping track
of previous logs and comparison in case of failure. Logs are collecting from
all the nodes in the cluster.
- The `post_fail_hook` will destroy all the defined barrier before collect any
logs. This action needs to occur in either of the modules, in order to make
the depended nodes available for log collection despite of which node is
failing and triggering the `post_fail_hook`.
- Provide option to pass `log_name` parameter via `upload_service_log` to
change the log file

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>