lib/y2_installbase: Allow more time for package searching
The select_packages test starts a search and then immediately calls
toggle_package. This leads to the issue that before the search is completed
it already presses the down key to move through the result list. Previously,
send_key_until_needlematch passed the timeout to the initial check_screen
call, which effectively gave it more time for the search to complete.

While this could probably be optimized by combining wait_still_screen with
a shorter check_screen, just reintroduce the previous behaviour for now.