Adjust hpc post_fail_hook for non-clustered nodes
Current commit introduced to collect logs from all nodes of a cluster. Logs
files would try to add the hostname of the node of each node. Clusters uses
`HOSTNAME` job variable which isnt assigned on a single machine jobs. Adjust
the post_fail_hook to use the default expected value.

Another solution would be to add `HOSTNAME` to each job which the variable is
missing such as the variable gives a meaningful name to the test logs.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>