Implement hpc modules to test dolly package in a cluster
Create a couple of files and copy them to other nodes using dolly. For more
info take a look at the documentation
https://documentation.suse.com/sle-hpc/15-SP3/html/hpc-guide/cha-dolly.html#
Many options are not supported so we go with the very basic scenario.

Some other approach/test cases might be to do the same with a config file.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>