wicked-wlan: delay namespace setup of wlan device
Give the system  some time to create the devices and throw all events.
This avoid a error message in wicked like
  Aug 25 02:35:16 install wickedd-nanny[646]: /org/opensuse/Network/Interface/4.getManagedObjects failed. Server responds:
  Aug 25 02:35:16 install wickedd-nanny[646]: org.freedesktop.DBus.Error.UnknownMethod: Method "GetManagedObjects" with signature "" on interface
which occur, because wickedd gets the NEWLINK and udev ADD event and emit a DEVICE_CREATED to nanny.
Nanny then tries to call the `getManagedObjects()` call, but the interface was already removed from wickedd again, because of the
move into the namespace.
Alternatively to `sleep` we could use a
   while ! dbus-send --system --print-reply --dest=org.opensuse.Network.Nanny /org/opensuse/Network/Nanny org.opensuse.Network.Nanny.getDevice string:"wlan0"; do sleep 1; done
but this would request `dbus-send` and wicked is running!