Optionally get Trento deployment scripts from a release tag
Add lib function to take care to get or just update the Trento
deployment script accordingly to the presence of TRENTO_DEPLOY_VER
variable.

Add lib function to get the public IP of the deployed Trento web app: it
also allow to specify the IP using a variable, meaning that the test has
to be performed against an instance of Trento deployed outside openQA.

Lib function to get the Trento web admin password.

Manage the version of the cypress/included docker image in a more
centralized way.

Remove most of the configuration variables from the schedule yaml file
as managed at JobGroup level or using default values in the test code.

Improved deployment scripts log management, expecially in case of
failure.